backend/pyproject.toml
[project]
name = "kgcfp-backend"
version = "0.1.0"
description = "FastAPI backend for KGCFP Knowledge Graph"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "neo4j>=6.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = []
